Fix issue with exporting adjunctions and pullbacks
We need to be able to determine angles of arrows, which relies on the size of columns and rows, which is stored in UI. Therefore, we now pass the UI into the export function. In the future, it would make sense to have this agnostic of the interface, but this will probably be subsumed by the arrow rewrite, so this is an acceptable fix for now.